mlnx-tools: Add ifup-local for roce_config
Trigger roce_config on ifup/ifdown by automatically installing
/sbin/ifup-local. Previously, a user had to manually create the
ifup-local script.

In the future this functionality will be provided by init scripts in
rdma-core so as to be vendor-independent.

For this interim version, the following assumptions are in place:

  1. there is no exsiting /sbin/ifup-local file (this is true of Oracle
     Linux)

  2. nobody else will be writing to this file so the only contents are
     those included here

What this means is that the mlnx-tools RPM will isntall /sbin/ifup-local
and will remove it if/when the RPM is uninstalled. If item 2 is
incorrect that could potentially remove other peoples' changes.

If the assumptions above change before this functionality is moved to
rdma-core, we can revisit this implementation.

This package in only supported on EL7. The ifup-local script is
installed using the macro %{_sbindir} which resolves to /usr/sbin even
though the script should be in /sbin. This works for EL7 because /sbin
is a symbolic link to /usr/sbin. On EL6 this would put the script in
/usr/sbin instead of /sbin and it would not be executed by initscripts.

Orabug: 26364780

Reviewed-by: Gerd Rausch <gerd.rausch@oracle.com>
Tested-by: Gerald Gibson <gerald.gibson@oracle.com>
Signed-off-by: Aron Silverton <aron.silverton@oracle.com>