ifup-local: Disable running roce_config by default
udev rules and a corresponding systemd service unit have been delivered to
Linux QA and Exadata. The udev rules tell systemd to schedule a service to do
the "roce configuration" and will be included in the Exadata image. The
configuration is only invoked for the interfaces that match the udev
rules and this is restricted to the RoCE interfaces only.

As such, the contents of ifup-local have been commented out as they are
no longer needed and could cause a conflict if the systems is
misconfigured and includes both /etc/systemd/network/re<n>.network
(correct) and /etc/sysconfig/network-scripts/ifcfg-re<n> (incorrect).

The file will be left in oracle-rdma-tools for now so that people
working on older systems without the Exadata-specific udev and systemd
enhancments can uncomment it and continue to work with SysV initscripts.

At some point in the future, the ifup-local file will be removed.

Orabug: 28122139

Signed-off-by: Aron Silverton <aron.silverton@oracle.com>