roce_config: Fix NETDEV matching and parsing
The output of 'ibdev2netdev' is matched and parsed to find the RDMA
device and port corresponding to a network device. Previously the
matching assumed that the NETDEV string is not part of the RDMA device
name. For example, searchign for 're0' in:

  mlx4_0 port 1 ==> ib0 (Down)
  mlx4_0 port 2 ==> ib1 (Down)
  mlx5_0 port 1 ==> re0 (Up)
  mlx5_0 port 2 ==> re1 (Up)

would only match in the fifth column, the NETDEV column, of the third
row. The corresponding RDMADEV from the first column or port from the
third column of that row would be returned.

However, if we have:

  bnxt_re0 port 1 ==> eth3 (Down)
  bnxt_re1 port 1 ==> eth4 (Down)
  mlx4_0 port 1 ==> ib0 (Down)
  mlx4_0 port 2 ==> ib1 (Down)
  mlx5_0 port 1 ==> re0 (Up)
  mlx5_0 port 2 ==> re1 (Up)

then we inadvertently match the RDMADEV in the first column of the first
row when we really want the NETDEV in the fifth row.

Orabug: 29030625

Signed-off-by: Aron Silverton <aron.silverton@oracle.com>
Reviewed-by: Gerald Gibson <gerald.gibson@oracle.com>