roce_config: Fix setting congestion control parameters
The script was attempting to set the RoCE congestion control priority
mask and RoCE congestion control algorithm for both port 1 and port
2 each time it configured an interface and failing under secure boot.

N.b: Congestion control was always set correctly. The FW sets it and the
     current script values do not override the FW settings.

This change makes it so that when configuring re0 it only attempts to
configure congestion control for port 1 and when configuring re1 it
only attempts to configure congestion control for port 2.

That is, only attempt to set parrameters for the physical port that
corresponds to the current NETDEV.

A hack is required to always use the first PCI function for the device
when using mstconfig to query or make the settings. This works because
the HCA's FW settings, even those only affecting one port, are
associated with the device itself and not the individual ports. We
probably need this hack because of the hack that turned the single-PF
dual-port HCA into something that looks like a dual-PF each with
a single port device.

Orabug: 29205754

Signed-off-by: Aron Silverton <aron.silverton@oracle.com>