roce-config: Update the host ToS map again
ToS 6 on the host was mapped to DSCP 44 (44 << 2 = 176) which is
priority queue 6 on the switch. However, DSCP 44 was already used for
priority 5:

    prio:5 dscp:47,46,45,44,43,42,41,40

As a result, traffic leaving the host with DSCP 44 actually flowed
through queue 5 on the switch not queue 6 mixing "small" and "large"
packets in the same queue.

The queue 6 packets are now multiplexed with queue 3 which is also used
for "large" messages. Queue 6 is currently not used for host-initiated
traffic (i.e., DSCP 44 is not mapped at the host.)

The resulting mappings are:

-----------------------------------------
CoS      DSCP    ToS Map  Switch Queue
-----------------------------------------
 1     8 ->  32    0,1    normal-large-q1
 2    16 ->  64    2,7    vip-small-q2
 3    24 ->  96    3,6    vip-large-q3
 4    32 -> 128    4      normal-small-q4
 5    40 -> 160    5      vip-small-q5
 6    44 -> 176           vip-large-q6
 7    56 -> 224           CNP

Fixes 613b92f4963e ("roce_config: Update the host ToS map")

Orabug: 29922090

Signed-off-by: Santosh Shilimkar <santosh.shilimkar@oracle.com>
Acked-by: Aron Silverton <aron.silverton@oracle.com>