ifup-local: Configure specific interfaces for RoCE
RoCE configuration (roce_config) will only be attempted for specific
interfaces. The interfaces may be specified as a list to match exactly
or as a pattern. Both are specified in a new configuration file
that is installed as "/etc/rdma/roce-interfaces.conf".

A list of interfaces may be something like: "eno2 enp94s0f0 enp94s0f1".
The list can contain a single entry or multiple entries (space
separated.) The list may also be empty or unset.

A pattern is defined by an interface base name (e.g., eno, eth, re)
and will match the RegEx "^<base name>\d+?". That is, the base name
followed by one or more digits. For examle, a base name of "re" will
match interfacs re0...re<n>. The base name "re" is the default and will
match RoCE interfaces for Exadata systems.

Orabug: 28122139

Signed-off-by: Aron Silverton <aron.silverton@oracle.com>
Tested-by: Gerald Gibson <gerald.gibson@oracle.com>